+++ Modules/SharePointDsc/DSCResources/MSFT_SPUserProfileSyncConnection/MSFT_SPUserProfileSyncConnection.psm1
@@ -424,5 +424,4 @@ function Get-SPDSCADSIObject
     return [ADSI]($LdapPath)
 }
 
-
 Export-ModuleMember -Function *-TargetResource, Get-SPDSCADSIObject