+++ Tests/Unit/SharePointDsc/SharePointDsc.SPManagedMetadataServiceApp.Tests.ps1
@@ -1656,4 +1656,5 @@ Describe -Name $Global:SPDscHelper.DescribeHeader -Fixture {
         }
     }
 }
+
 Invoke-Command -ScriptBlock $Global:SPDscHelper.CleanupScript -NoNewScope